requirements.txt
# Dream.os - Core Dependencies
# Generated from pyproject.toml for CI/CD compatibility

# Runtime dependencies (required for code execution)
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
aiohttp>=3.8.0
<<<<<<< HEAD
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
discord.py>=2.3.0
pyautogui>=0.9.54
pyperclip>=1.8.0
=======
discord.py>=2.3.0
pyautogui>=0.9.54
>>>>>>> origin/codex/build-cross-platform-control-plane-for-swarm-console

# Code formatting and linting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
autopep8>=2.0.0
pre-commit>=3.0.0
pylint>=2.17.0
mypy>=1.0.0
bandit>=1.7.0
safety>=2.0.0
ruff==0.6.9
pyupgrade==3.15.2
bowler==0.9.0

# Testing (included for CI)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0  # Parallel test execution
pytest-html>=3.1.0  # HTML test reports
